clarify the function of Experiment and ControlPanel. Complete the pydoc of those.
